+++ builder/parallels/common/prlctl_post_config.go
@@ -1,3 +1,5 @@
+//go:generate struct-markdown
+
 package common
 
 import (
@@ -7,7 +9,10 @@ import (
 // PrlctlPostConfig contains the configuration for running "prlctl" commands
 // in the end of artifact build.
 type PrlctlPostConfig struct {
-	PrlctlPost [][]string `mapstructure:"prlctl_post"`
+	// Identical to prlctl, except
+	// that it is run after the virtual machine is shutdown, and before the virtual
+	// machine is exported.
+	PrlctlPost [][]string `mapstructure:"prlctl_post" required:"false"`
 }
 
 // Prepare sets the default value of "PrlctlPost" property.