+++ builder/parallels/common/prlctl_version_config_test.go
@@ -2,6 +2,8 @@ package common
 
 import (
 	"testing"
+
+	"github.com/hashicorp/packer/template/interpolate"
 )
 
 func TestPrlctlVersionConfigPrepare_BootWait(t *testing.T) {
@@ -10,7 +12,7 @@ func TestPrlctlVersionConfigPrepare_BootWait(t *testing.T) {
 
 	// Test empty
 	c = new(PrlctlVersionConfig)
-	errs = c.Prepare(testConfigTemplate(t))
+	errs = c.Prepare(interpolate.NewContext())
 	if len(errs) > 0 {
 		t.Fatalf("should not have error: %s", errs)
 	}
@@ -22,7 +24,7 @@ func TestPrlctlVersionConfigPrepare_BootWait(t *testing.T) {
 	// Test with a good one
 	c = new(PrlctlVersionConfig)
 	c.PrlctlVersionFile = "foo"
-	errs = c.Prepare(testConfigTemplate(t))
+	errs = c.Prepare(interpolate.NewContext())
 	if len(errs) > 0 {
 		t.Fatalf("should not have error: %s", errs)
 	}