+++ communicator/ssh/communicator.go
@@ -35,6 +35,24 @@ type comm struct {
 	address string
 }
 
+// TunnelDirection is the supported tunnel directions
+type TunnelDirection int
+
+const (
+	UnsetTunnel TunnelDirection = iota
+	RemoteTunnel
+	LocalTunnel
+)
+
+// TunnelSpec represents a request to map a port on one side of the SSH connection to the other
+type TunnelSpec struct {
+	Direction   TunnelDirection
+	ListenType  string
+	ListenAddr  string
+	ForwardType string
+	ForwardAddr string
+}
+
 // Config is the structure used to configure the SSH communicator.
 type Config struct {
 	// The configuration of the Go SSH connection
@@ -64,6 +82,8 @@ type Config struct {
 
 	// Timeout is how long to wait for a read or write to succeed.
 	Timeout time.Duration
+
+	Tunnels []TunnelSpec
 }
 
 // Creates a new packer.Communicator implementation over SSH. This takes
@@ -344,8 +364,74 @@ func (c *comm) reconnect() (err error) {
 		c.client = ssh.NewClient(sshConn, sshChan, req)
 	}
 	c.connectToAgent()
+	err = c.connectTunnels(sshConn)
+	if err != nil {
+		return
+	}
 
 	return
+}
+
+func (c *comm) connectTunnels(sshConn ssh.Conn) (err error) {
+	if c.client == nil {
+		return
+	}
+
+	if len(c.config.Tunnels) == 0 {
+		// No Tunnels to configure
+		return
+	}
+
+	// Start remote forwards of ports to ourselves.
+	log.Printf("[DEBUG] Tunnel configuration: %v", c.config.Tunnels)
+	for _, v := range c.config.Tunnels {
+		done := make(chan struct{})
+		var listener net.Listener
+		switch v.Direction {
+		case RemoteTunnel:
+			// This requests the sshd Host to bind a port and send traffic back to us
+			listener, err = c.client.Listen(v.ListenType, v.ListenAddr)
+			if err != nil {
+				err = fmt.Errorf("Tunnel: Failed to bind remote ('%v'): %s", v, err)
+				return
+			}
+			log.Printf("[INFO] Tunnel: Remote bound on %s forwarding to %s", v.ListenAddr, v.ForwardAddr)
+			connectFunc := ConnectFunc(v.ForwardType, v.ForwardAddr)
+			go ProxyServe(listener, done, connectFunc)
+			// Wait for our sshConn to be shutdown
+			// FIXME: Is there a better "on-shutdown" we can wait on?
+			go shutdownProxyTunnel(sshConn, done, listener)
+		case LocalTunnel:
+			// This binds locally and sends traffic back to the sshd host
+			listener, err = net.Listen(v.ListenType, v.ListenAddr)
+			if err != nil {
+				err = fmt.Errorf("Tunnel: Failed to bind local ('%v'): %s", v, err)
+				return
+			}
+			log.Printf("[INFO] Tunnel: Local bound on %s forwarding to %s", v.ListenAddr, v.ForwardAddr)
+			connectFunc := func() (net.Conn, error) {
+				// This Dial occurs on the SSH server's side
+				return c.client.Dial(v.ForwardType, v.ForwardAddr)
+			}
+			go ProxyServe(listener, done, connectFunc)
+			// FIXME: Is there a better "on-shutdown" we can wait on?
+			go shutdownProxyTunnel(sshConn, done, listener)
+		default:
+			err = fmt.Errorf("Tunnel: Unknown tunnel direction ('%v'): %v", v, v.Direction)
+			return
+		}
+	}
+
+	return
+}
+
+// shutdownProxyTunnel waits for our sshConn to be shutdown and closes the listeners
+func shutdownProxyTunnel(sshConn ssh.Conn, done chan struct{}, listener net.Listener) {
+	sshConn.Wait()
+	log.Printf("[INFO] Tunnel: Shutting down listener %v", listener)
+	done <- struct{}{}
+	close(done)
+	listener.Close()
 }
 
 func (c *comm) connectToAgent() {